+++ src/components/sections/elements/Scale.vue
@@ -1,0 +1,58 @@
+<script setup lang="ts">
+import { computed, type ComputedRef, onMounted, ref, watch } from 'vue'
+
+import { Stability, SectionType } from '@/types/enum'
+
+import SectionBorder from '@/components/bases/SectionBorder.vue'
+import SectionTitle from '@/components/bases/SectionTitle.vue'
+import Output from '@/components/bases/Output.vue'
+import FormInput from '@/components/bases/FormInput.vue'
+import TwoColumn from '@/components/bases/TwoColumn.vue'
+import GuidanceText from '@/components/bases/GuidanceText.vue'
+
+const emit = defineEmits<{
+  'update:isoSpatialResolution': [id: number]
+}>()
+
+const scale = ref<number>(0)
+
+let scale_fmt: ComputedRef<string> = computed(() => {
+  return `1:${scale.value.toLocaleString()}`
+})
+
+onMounted(() => {
+  emit('update:isoSpatialResolution', scale.value)
+})
+
+watch(
+  () => scale.value,
+  () => {
+    emit('update:isoSpatialResolution', scale.value)
+  }
+)
+</script>
+
+<template>
+  <SectionBorder :type="SectionType.Element">
+    <SectionTitle
+      :stability="Stability.Experimental"
+      :type="SectionType.Element"
+      version="1.0"
+      anchor="scale"
+      title="Scale"
+    />
+    <TwoColumn>
+      <template #left>
+        <div>
+          <FormInput id="scale" v-model="scale" type="number" name="scale" class="w-full" />
+        </div>
+      </template>
+      <template #right>
+        <div class="space-y-2">
+          <Output :data="scale"></Output>
+          <GuidanceText>Read as '{{ scale_fmt }}'.</GuidanceText>
+        </div>
+      </template>
+    </TwoColumn>
+  </SectionBorder>
+</template>